refactor: modal to overlay
