+++ ui/overlay/modal_test.go
@@ -1,4 +1,4 @@
-package modal
+package overlay
 
 import (
 	"bytes"
